+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed an issue with `simple-bus` binding not found when using zephyr bindings.
 - Fixed an issue where external file change events via the file watchers where ignored
   if the file is open and saved in the editor.
+- Fixed zephyr binding const validation did not work when expression where used.
 
 ## [0.7.8] - 2026-01-25